# docs/api/management-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GEOFlow remote management preview",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/auth/session": {
            "get": {
                "operationId": "auth.session",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "x-required-scope": null,
                "x-idempotent": false,
                "x-receipt": false,
                "responses": {
                    "200": {
                        "description": "Successful API envelope"
                    },
                    "401": {
                        "description": "Authentication required"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "409": {
                        "description": "Baseline or request identity conflict"
                    }
                }
            }
        },
...
        "/management/theme-workspaces": {
            "post": {
                "operationId": "theme-workspaces.create",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "x-required-scope": "themes:write",
                "x-idempotent": false,
                "x-receipt": false,
                "responses": {
                    "200": {
                        "description": "Successful API envelope"
                    },
                    "401": {
                        "description": "Authentication required"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "409": {
                        "description": "Baseline or request identity conflict"
                    },
                    "201": {
                        "description": "Resource or receipt created"
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "site",
                                    "theme"
                                ]
                            }
                        }
                    }
                }
            }
        },
...
        "/management/theme-workspaces/{workspace}/previews": {
            "post": {
                "operationId": "theme-workspaces.preview",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "x-required-scope": "themes:read",
                "x-idempotent": false,
                "x-receipt": false,
                "responses": {
                    "200": {
                        "description": "Successful API envelope"
                    },
                    "401": {
                        "description": "Authentication required"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "409": {
                        "description": "Baseline or request identity conflict"
                    },
                    "201": {
                        "description": "Resource or receipt created"
                    }
                },
                "parameters": [
                    {
                        "name": "workspace",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/management/theme-workspaces/{workspace}": {
            "get": {
                "operationId": "theme-workspaces.show",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "x-required-scope": "themes:read",
                "x-idempotent": false,
                "x-receipt": false,
                "responses": {
                    "200": {
                        "description": "Successful API envelope"
                    },
                    "401": {
                        "description": "Authentication required"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "409": {
                        "description": "Baseline or request identity conflict"
                    }
                },
                "parameters": [
                    {
                        "name": "workspace",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
        "/management/themes": {
            "get": {
                "operationId": "themes.list",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "x-required-scope": "themes:read",
                "x-idempotent": false,
                "x-receipt": false,
                "responses": {
                    "200": {
                        "description": "Successful API envelope"
                    },
                    "401": {
                        "description": "Authentication required"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "409": {
                        "description": "Baseline or request identity conflict"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}